requirements.txt
streamlit>=1.28.0,<2
streamlit-webrtc
av
opencv-python-headless>=4.8.0,<5
ultralytics>=8.3.0,<9
clip @ git+https://github.com/ultralytics/CLIP.git
deepface>=0.0.90,<1
tf-keras>=2.15.0,<3
pillow>=10.0.0,<12
pandas>=2.0.0,<3
numpy>=1.24.0,<3
python-dotenv>=1.0.0,<2
bcrypt>=4.1.0,<5
cryptography>=42.0.0,<46
google-auth>=2.0.0,<3
google-auth-oauthlib>=1.0.0,<2
google-auth-httplib2>=0.2.0,<1
email-validator>=2.0.0,<3
plotly>=5.18.0,<6
scikit-learn>=1.3.0,<2
pydantic>=2.0.0,<3
requests>=2.31.0,<3
pytest>=8.0.0,<10
fpdf2>=2.7.0,<3
pyotp>=2.9.0,<3
qrcode>=7.4.0,<9
